+++ src/PFConst.js
@@ -39,4 +39,4 @@ export default {
 	minusreg : /\-|\u2013|\u2014|\u2212|\u02d7/,
 	critreg : /(\d+)[\-|\u2013|\u2014|\u2212|\u02d7]20\/[x\u00d7](\d+)/,
 	diceDiereg : /(\d+)d(\d+)\s*([\+|\-|\u2013|\u2014|\u2212|\u02d7]{0,1})\s*(\d*)/
-}
+};
