+++ src/PFLog.js
@@ -1,3 +1,4 @@
+'use strict';
 export const PFLog = {
  // PFLog just a little helper to write to console using fancy type
  l: '%cס§₪₪₪₪§|(Ξ≥≤≥≤≥≤ΞΞΞΞΞΞΞΞΞΞ>    ',
@@ -10,4 +11,4 @@ export const PFConsole = {
   log: (text) => {
     console.log(PFLog.l + text + PFLog.r, PFLog.bg);
   }
-}
+};
